Adiciona melhorias

+++ Domain/Models/DadosAtraso.cs
@@ -2,9 +2,9 @@ namespace contas.api.Domain.Models
 {
     public class DadosAtraso
     {
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
         public int DiasAtraso { get; set; }
-        public double Multa { get; set; }
-        public double Juros { get; set; }
+        public decimal Multa { get; set; }
+        public decimal Juros { get; set; }
     }
 }
